move ctrl_util to util package
